requirements.txt
click==8.0.1
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.9.3
python-dotenv==1.0.0
SQLAlchemy==1.4.18
Werkzeug==2.0.1
